LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/var-reference-registered-properties-cycles-expected.txt


FAIL A var() cycle between two registered properties is handled correctly. assert_equals: expected "1px" but got ""
FAIL A var() cycle between a registered properties and an unregistered property is handled correctly. assert_equals: expected "1px" but got ""
FAIL A var() cycle between a two unregistered properties is handled correctly. assert_equals: expected "30px" but got ""
FAIL A var() cycle between a syntax:'*' property and an unregistered property is handled correctly. assert_equals: expected "meow" but got ""
